+++ src/main/java/ladysnake/blast/common/init/BlastItems.java
@@ -36,25 +36,25 @@ public class BlastItems {
     public static Item FROST_TRIGGER_BOMB;
 
     public static void init() {
-        BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.TOOLS), BlastEntities.BOMB), "bomb");
-        TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.TOOLS), BlastEntities.TRIGGER_BOMB), "trigger_bomb");
-        GOLDEN_BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.TOOLS), BlastEntities.GOLDEN_BOMB), "golden_bomb");
-        GOLDEN_TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.TOOLS), BlastEntities.GOLDEN_TRIGGER_BOMB), "golden_trigger_bomb");
-        DIAMOND_BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.TOOLS), BlastEntities.DIAMOND_BOMB), "diamond_bomb");
-        DIAMOND_TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.TOOLS), BlastEntities.DIAMOND_TRIGGER_BOMB), "diamond_trigger_bomb");
-        NAVAL_MINE = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.TOOLS), BlastEntities.NAVAL_MINE), "naval_mine");
-        CONFETTI_BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.TOOLS), BlastEntities.CONFETTI_BOMB), "confetti_bomb");
-        CONFETTI_TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.TOOLS), BlastEntities.CONFETTI_TRIGGER_BOMB), "confetti_trigger_bomb");
-        DIRT_BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.TOOLS), BlastEntities.DIRT_BOMB), "dirt_bomb");
-        DIRT_TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.TOOLS), BlastEntities.DIRT_TRIGGER_BOMB), "dirt_trigger_bomb");
-        PEARL_BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.TOOLS), BlastEntities.PEARL_BOMB), "pearl_bomb");
-        PEARL_TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.TOOLS), BlastEntities.PEARL_TRIGGER_BOMB), "pearl_trigger_bomb");
-        SLIME_BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.TOOLS), BlastEntities.SLIME_BOMB), "slime_bomb");
-        SLIME_TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.TOOLS), BlastEntities.SLIME_TRIGGER_BOMB), "slime_trigger_bomb");
-        AMETHYST_BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.COMBAT), BlastEntities.AMETHYST_BOMB), "amethyst_bomb");
-        AMETHYST_TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.COMBAT), BlastEntities.AMETHYST_TRIGGER_BOMB), "amethyst_trigger_bomb");
-        FROST_BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.COMBAT), BlastEntities.FROST_BOMB), "frost_bomb");
-        FROST_TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.COMBAT), BlastEntities.FROST_TRIGGER_BOMB), "frost_trigger_bomb");
+        BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.TOOLS).maxCount(16), BlastEntities.BOMB), "bomb");
+        TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.TOOLS).maxCount(16), BlastEntities.TRIGGER_BOMB), "trigger_bomb");
+        GOLDEN_BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.TOOLS).maxCount(16), BlastEntities.GOLDEN_BOMB), "golden_bomb");
+        GOLDEN_TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.TOOLS).maxCount(16), BlastEntities.GOLDEN_TRIGGER_BOMB), "golden_trigger_bomb");
+        DIAMOND_BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.TOOLS).maxCount(16), BlastEntities.DIAMOND_BOMB), "diamond_bomb");
+        DIAMOND_TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.TOOLS).maxCount(16), BlastEntities.DIAMOND_TRIGGER_BOMB), "diamond_trigger_bomb");
+        NAVAL_MINE = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.TOOLS).maxCount(16), BlastEntities.NAVAL_MINE), "naval_mine");
+        CONFETTI_BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.TOOLS).maxCount(16), BlastEntities.CONFETTI_BOMB), "confetti_bomb");
+        CONFETTI_TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.TOOLS).maxCount(16), BlastEntities.CONFETTI_TRIGGER_BOMB), "confetti_trigger_bomb");
+        DIRT_BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.TOOLS).maxCount(16), BlastEntities.DIRT_BOMB), "dirt_bomb");
+        DIRT_TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.TOOLS).maxCount(16), BlastEntities.DIRT_TRIGGER_BOMB), "dirt_trigger_bomb");
+        PEARL_BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.TOOLS).maxCount(16), BlastEntities.PEARL_BOMB), "pearl_bomb");
+        PEARL_TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.TOOLS).maxCount(16), BlastEntities.PEARL_TRIGGER_BOMB), "pearl_trigger_bomb");
+        SLIME_BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.TOOLS).maxCount(16), BlastEntities.SLIME_BOMB), "slime_bomb");
+        SLIME_TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.TOOLS).maxCount(16), BlastEntities.SLIME_TRIGGER_BOMB), "slime_trigger_bomb");
+        AMETHYST_BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.COMBAT).maxCount(16), BlastEntities.AMETHYST_BOMB), "amethyst_bomb");
+        AMETHYST_TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.COMBAT).maxCount(16), BlastEntities.AMETHYST_TRIGGER_BOMB), "amethyst_trigger_bomb");
+        FROST_BOMB = registerItem(new BombItem(new Item.Settings().group(ItemGroup.COMBAT).maxCount(16), BlastEntities.FROST_BOMB), "frost_bomb");
+        FROST_TRIGGER_BOMB = registerItem(new TriggerBombItem(new Item.Settings().group(ItemGroup.COMBAT).maxCount(16), BlastEntities.FROST_TRIGGER_BOMB), "frost_trigger_bomb");
     }
 
     public static Item registerItem(Item item, String name) {